add try n catch

+++ src/controllers/userController.js
@@ -1,73 +1,68 @@
 const jwt = require("jsonwebtoken");
 const userModel = require("../models/userModel");
 
-const createUser = async function (abcd, xyz) {
-  //You can name the req, res objects anything.
-  //but the first parameter is always the request 
-  //the second parameter is always the response
-  let data = abcd.body;
-  let savedData = await userModel.create(data);
-  console.log(abcd.newAtribute);
-  xyz.send({ msg: savedData });
+const createUser = async function (req, res) {
+  try {
+    let data = req.body;
+    if (Object.keys(data).length != 0) {
+      let savedData = await userModel.create(data);
+      res.status(201).send({ status: true, data: savedData })
+    } else res.status(400).send({ status: false, msg: "BAD REQUEST" })
+  }
+  catch (error) {
+    res.status(500).send({ status: false, msg: error.message })
+  }
 };
+//================================================================================================
 
 const loginUser = async function (req, res) {
-  let userName = req.body.emailId;
-  let password = req.body.password;
+  try {
+    if (req.body && req.body.emailId && req.body.password) {
 
-  let user = await userModel.findOne({ emailId: userName, password: password });
-  if (!user)
-    return res.send({
-      status: false,
-      msg: "username or the password is not corerct",
-    });
-
-  // Once the login is successful, create the jwt token with sign function
-  // Sign function has 2 inputs:
-  // Input 1 is the payload or the object containing data to be set in token
-  // The decision about what data to put in token depends on the business requirement
-  // Input 2 is the secret
-  // The same secret will be used to decode tokens
-  let token = jwt.sign(
-    {
-      userId: user._id.toString(),
-      batch: "thorium",
-      organisation: "FUnctionUp",
-    },
-    "functionup-thorium"
-  );
-  res.setHeader("x-auth-token", token);
-  res.send({ status: true, data: token });
+      let user = await userModel.findOne({ emailId: req.body.emailId, password: req.body.password });
+      if (user) {
+        let token = jwt.sign(
+          {
+            userId: user._id.toString(),
+            batch: "lithium",
+            organisation: "FUnctionUp",
+          },
+          "functionup-lithium"
+        );
+        res.setHeader("x-auth-token", token);
+        res.status(200).send({ status: true, data: token });
+      } else {
+        res.status(401).send({ status: false, msg: "invalid emailId or password" })
+      }
+    } else {
+      res.status(400).send({ status: false, msg: "request body must be contain emailId and password" })
+    }
+  } catch (error) {
+    res.status(500).send({ status: false, msg: error.message })
+  }
 };
+//========================================================================================================
 
 const getUserData = async function (req, res) {
- 
-  //If no token is present in the request header return error
- 
-  
-  // If a token is present then decode the token with verify function
-  // verify takes two inputs:
-  // Input 1 is the token to be decoded
-  // Input 2 is the same secret with which the token was generated
-  // Check the value of the decoded token yourself
-  
-  let userId = req.params.userId;
-  let userDetails = await userModel.findById(userId);
-  if (!userDetails)
-    return res.send({ status: false, msg: "No such user exists" });
+  try {
 
-  res.send({ status: true, data: userDetails });
+    let userId = req.params.userId;
+    let userDetails = await userModel.findById(userId);
+    if (userDetails) {
+      res.status(200).send({ status: true, data: userDetails })
+    } else {
+      res.status(404).send({ status: false, msg: "user not found" })
+    }
+  } catch (error) {
+    res.status(500).send({ status: false, msg: error.message })
+  }
 };
+//============================================================================================
 
 const updateUser = async function (req, res) {
-// Do the same steps here:
-// Check if the token is present
-// Check if the token present is a valid token
-// Return a different error message in both these cases
 
   let userId = req.params.userId;
   let user = await userModel.findById(userId);
-  //Return an error if no user with the given id exists in the db
   if (!user) {
     return res.send("No such user exists");
   }
@@ -76,39 +71,43 @@ const updateUser = async function (req, res) {
   let updatedUser = await userModel.findOneAndUpdate({ _id: userId }, userData);
   res.send({ status: updatedUser, data: updatedUser });
 };
+//=============================================================================================
 
 const postMessage = async function (req, res) {
-    let message = req.body.message
-    // Check if the token is present
-    // Check if the token present is a valid token
-    // Return a different error message in both these cases
-   
-    //userId for which the request is made. In this case message to be posted.
-    let userToBeModified = req.params.userId
-    //userId for the logged-in user
-    let userLoggedIn = decodedToken.userId
+  try{
+  let message = req.body.message
+  let userToBeModified = req.params.userId
+  let userLoggedIn = decodedToken.userId
 
-    //userId comparision to check if the logged-in user is requesting for their own data
-    if(userToBeModified != userLoggedIn) return res.send({status: false, msg: 'User logged is not allowed to modify the requested users data'})
+  if (userToBeModified != userLoggedIn) return res.send({ status: false, msg: 'User logged is not allowed to modify the requested users data' })
 
-    let user = await userModel.findById(req.params.userId)
-    if(!user) return res.send({status: false, msg: 'No such user exists'})
-    
-    let updatedPosts = user.posts
-    //add the message to user's posts
-    updatedPosts.push(message)
-    let updatedUser = await userModel.findOneAndUpdate({_id: user._id},{posts: updatedPosts}, {new: true})
+  let user = await userModel.findById(req.params.userId)
+  if (!user) return res.send({ status: false, msg: 'No such user exists' })
 
-    //return the updated user document
-    return res.send({status: true, data: updatedUser})
+  let updatedPosts = user.posts
+  updatedPosts.push(message)
+  let updatedUser = await userModel.findOneAndUpdate({ _id: user._id }, { posts: updatedPosts }, { new: true })
+
+  return res.send({ status: true, data: updatedUser })
+  } catch (error) {
+    res.status(500).send({ status: false, msg: error.message })
+  }
 }
+
+//================================================================================================
+
 const deleteUser = async function (req, res) {
-  let dlt = req.params.userId
- let isDeleted = await userModel.findByIdAndDelete(dlt);
- res.send({ status:"Deleted", updatedData: isDeleted });
-
+  try {
+    let dlt = req.params.userId
+    if (dlt === userId) {
+      let isDeleted = await userModel.findByIdAndDelete(dlt);
+      res.status(200).send({ status: "Deleted", updatedData: isDeleted });
+    } else res.status(404).send({ status: false, msg: "USER NOT FOUND" })
+  } catch (error) {
+    res.status(500).send({ status: false, msg: error.message })
+  }
 };
-
+//==================================================================================================
 
 module.exports.createUser = createUser;
 module.exports.getUserData = getUserData;
